requirements.txt
torch torchvision
git+https://github.com/openai/CLIP.git
pillow
numpy
faiss-cpu
streamlit
scikit-learn
tqdm
exifread
face_recognition
